Added Average of Array

+++ Arrays/AverageOfArray.java
@@ -6,7 +6,7 @@ public class AverageOfArray {
         for(int i=0; i<n; i++){
             sum += arr[i];
         }
-        int avg = sum/n;
+        double avg = (double)sum/n;
         System.out.println("Average of all elements is :" + avg);
     }
 }
